+++ R/fit_topt_VJs.R
@@ -62,8 +62,6 @@
 #' #Let's get the graphs out into a list
 #' #You can get a graph using: graph[1]
 #' graphs <- get_t_graphs(out)
-#' #Print graphs out as jpegs into folder
-#' print_graphs(graphs)
 #' }
 fit_topt_VJs <- function(data,
                          group,