+++ R/print_graphs.R
@@ -1,11 +1,16 @@
 #' Printing graphs from a list of graphs
 #'
 #' @param data List of graphs to output as .jpeg files
+#' @param height Height of output graphs
+#' @param width Width of output graphs
+#' @param res Resolution of output graphs
+#' @param units Units of height and width
 #' @return print_graphs creates jpeg files from a list of graphs based on
 #' the graph names. Used in combination with get_t_graphs. Output is a
 #' series of .jpeg files in the working directory.
 #' @importFrom grDevices dev.off
 #' @importFrom grDevices jpeg
+#' @importFrom utils prompt
 #' @export
 #' @examples \donttest{
 #' #Read in data
@@ -47,11 +52,25 @@
 #' #Let's get the graphs out into a list
 #' #You can get a graph using: graph[1]
 #' graphs <- get_t_graphs(out)
+#' #Set temporary working directory for example
+#' setwd(tempdir())
 #' #Print graphs out as jpegs into folder
-#' #Uncomment to print graphs
-#' #print_graphs(graphs)
+#' print_graphs(graphs)
+#' print("N")
 #' }
-print_graphs <- function(data) {
+print_graphs <- function(data,
+                         height = 5,
+                         width = 5,
+                         res = 600,
+                         units = "in") {
+  #Show user working directory
+  print(getwd())
+  #Ask user for permission to print the graphs to working directory
+  user.prompt <- readline(prompt = 
+  "Do you want to print graphs to your current working directory? Y/N: ")
+  #When permission is given, print graphs to working directory
+  #Otherwise say graphs were not printed
+  if (user.prompt == "Y") {
   #Print out each graph in the list to the current working directory
   for (i in 1:length(data)) {
     jpeg(paste(names(data[[i]])[1], ".jpeg"),
@@ -59,5 +78,8 @@ print_graphs <- function(data) {
          units = "in")
     print(data[[i]])
     dev.off()
+  }
+  } else {
+    print("Graphs not printed")
   }
 }